Enhancement #35 (requiring migration)
Posts on the index are now ordered by post and comment activity. Deleted comments update the post activity. Post page functon was updated. An action in the admin panel was added to allow posts modified date to be refreshed, if they are out of date or do not match correctly post migration.

This update requires a database migration.